+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
     Currently there two side-by-side build systems that produces the same output (net40) from the same sources.
     After a some transition time, current (mono/ msbuild 14.0) build system will be removed.
 -   NUnit upgraded to 3.7 (eliminates travis-ci random bug)
+-   Added C# `PythonEngine.AddShutdownHandler` to help client code clean up on shutdown.
 -   Added `clr.GetClrType` ([#432][i432])([#433][p433])
 -   Allowed passing `None` for nullable args ([#460][p460])
 -   Added keyword arguments based on C# syntax for calling CPython methods ([#461][p461])
@@ -43,6 +44,7 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 -   Fixed PyObject.GetHashCode ([#676][i676])
 -   Fix memory leaks due to spurious handle incrementation ([#691][i691])
 -   Fix spurious assembly loading exceptions from private types ([#703][i703])
+-   Fix inheritance of non-abstract base methods ([#755][i755])
 
 
 ## [2.3.0][] - 2017-03-11
@@ -600,6 +602,7 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 
 [1.0.0]: https://github.com/pythonnet/pythonnet/releases/tag/1.0
 
+[i714]: https://github.com/pythonnet/pythonnet/issues/714
 [i608]: https://github.com/pythonnet/pythonnet/issues/608
 [i443]: https://github.com/pythonnet/pythonnet/issues/443
 [p690]: https://github.com/pythonnet/pythonnet/pull/690
@@ -693,3 +696,4 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 [p625]: https://github.com/pythonnet/pythonnet/pull/625
 [i131]: https://github.com/pythonnet/pythonnet/issues/131
 [p531]: https://github.com/pythonnet/pythonnet/pull/531
+[i755]: https://github.com/pythonnet/pythonnet/pull/755